backend/requirements.txt
# FastAPI core
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10
greenlet==3.3.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Pydantic & Settings (Python 3.13 compatible)
pydantic==2.10.4
pydantic-settings==2.7.0
email-validator==2.2.0

# Cloud Storage
cloudinary==1.41.0
boto3==1.35.99

# Redis caching (optional)
redis==5.2.1

# HTTP client
httpx==0.28.1

# CORS & Middleware
python-dotenv==1.0.1

# Date handling
python-dateutil==2.9.0.post0

# Logging
loguru==0.7.3
